requirements_dev.txt
-r requirements.txt
coverage
coveralls
docutils
flake8<7.0.0
flake8-quotes>=3.3.2
hatch
hatchling
jupyter
mypy>=1.16.0
nbconvert
nbval
pylint>=3.2.0
pytest
python-Levenshtein
scipy
sortedcontainers-stubs
phinx
twine
types-requests
wheel
